Basic tools for parsing SSH fields.
Motivation:

SSH has a number of specific data layouts that apply only to it, which
we should have helpers for. This will make our lives easier when parsing
packets down the road.

This code does not include a change for the `mpint` type, which is
fundamentally just a string and will be treated as such in code, or for
`name-list`. In the case of `name-list`, we will likely want to design a
solution for this at a future date, as we'll probably need a custom
iterator type at that stage.

Modifications:

- Added tools for reading SSH strings and booleans.

Result:

Easier parsing code.