AES GCM implementation
Motivation:

For compatibility with OpenSSH we'd like to support a cipher suite that
it knows how to speak. In the overlap of CryptoKit and OpenSSH that
basically boils down to just the AES-GCM protocols, so we're going to go
with those.

Modifications:

- Implement aes128-gcm@openssh.com and aes256-gcm@openssh.com
- Some basic testing.

Result:

We'll be able to encrypt data that OpenSSH can read.