Define CSPRNG for internal usage.
Motivation:

SSH has a number of places where we need access to a CSPRNG.
Unfortunately, Swift doesn't really give us a type that meets this
requirement. While we could build one ourselves, it would largely be a
carbon copy of the code that currently exists in the core Swift project.

For the sake of pragmatism, for now we shouldn't invest too much time in
writing this ourselves. Instead, we believe that
SystemRandomNumberGenerator is good and will remain good on a few
specific platforms, so we restrict its use to those platforms and fail
hard anywhere else.

Modifications:

- Added an internal CSPRNG type deferring to
    SystemRandomNumberGenerator.

Result:

There will be cryptographically secure randomness for SSH.