Add User Auth state machine.
Motivation:

When performing the user authentication part of an SSH connection
establishment we need to keep track of what we're doing and why. This
requires a state machine.

The User Authentication state machine will drive the connection during
the user auth phase. It allows plugging in of delegates to control how
authentication will work: one delegate for the client and one for the
server.

The current implementation is somewhat limited, but it's a good starting
point to get things plugged in and validate that the basic properties
work. To begin with, the user auth state machine does not allow the
client to perform parallel authentication attempts: the client may have
one auth request outstanding at any time. The server is not bound by
this because we don't control clients, so the server is required to
tolerate the possibility of receiving multiple auth requests at once.

Additionally, we are currently limited to password authentication only.
Before we ship I intend to widen the list to public key: however, we may
never truly support hostbased authentication. It depends on how
necessary it is.

This patch does not integrate the user authentication state machine into
the main connection state machine. That will come in a future patch.

Modifications:

- Implemented the UserAuthenticationStateMachine.
- Defined ClientUserAuthenticationDelegate.
- Defined ServerUserAuthenticationDelegate.
- Wrote a lot of tests.

Result:

We'll have a state machine we can use.