Update state machine interface and improve kex.
As the user authentication state machine requires the ability to perform
asynchronous operations, it is not tenable to have our state machine
synchronously generate messages and assume that they'll be immediately
sent. Instead, we need to have a multi-stage process where messages are
generated but have separate send steps.

As part of this work it became worthwhile to refactor the key exchange
state machine to better serve this use-case. This adds some complexity
to it but also unlocks some useful optimisations, such as the ability to
save an RTT when sending the initial KeyExchange message.