Add support for multiple host key types.
Motivation:

There are plenty of clients and servers that don't speak exactly the set
of host key formats that we do. For this reason, it's valuable for
servers to have multiple host keys, and for clients to have the ability
to tolerate multiple as well.

We never added support for this, but we should.

Modifications:

- Expanded key exchange to keep track of negotiated host key formats.

Result:

More flexibility in host key formats.