Better handling of TCP FIN during setup.
Motivation:

When setting up a child channel there were windows where TCP FIN would
cause us to hit an error case. Most notably this happened if we either
hadn't completed our initializer or we hadn't received a response to
CHANNEL_OPEN. In both cases, we should do better.

Additionally, we were piggybacking on some state machine logic even in
cases where the channel was setup in order to try to avoid duplicating
code. This led us to write messages that we knew were doomed to failure.
This patch takes this opportunity to fix that as well.

Modifications:

- Enhance the child channel state machine to know how to respond to TCP
  FIN.

Result:

Better behaviour on TCP FIN.