Add support for public key user auth.
Motivation:

Public key user auth is vital for a functioning SSH infrastructure, for
two reasons. Firstly, it's the most secure kind, avoiding the need to
transmit passwords across a network or force users to type them in. And
secondly, it's the easiest for computers to use, as it does not involve
finding awkward ways to community messy textual passwords around.

Modifications:

- Added parsing and handling support for SSH keys.
- Moved the host key types out into fully-fledged key types.
- Tolerated unknown public key types.
- Fixed a number of ECDSA bugs.

Result:

Public key user auth is real!