Define new configuration logic and global request delegate (#2)
Motivation:

As part of the work for TCP port forwarding support we need a way for
"global requests" to be handled. This implies the existence of a "Global
Request Delegate". When adding this it became clear that configuration
was getting out of hand. This patch rewrites configuration to make it a
bit easier to handle.

Modifications:

- Added new GlobalRequestDelegate
- Added SSHClientConfiguration and SSHServerConfiguration
- Moved existing delegate and host key setup under the new config.
- Fixed up the tests.

Results:

Single configuration object.