Add support for TCP port forwarding. (#4)
Motivation:

SSH provides for the valuable feature of remote and local TCP port
forwarding. This allows SSH clients and servers to arrange for TCP
connections to be tunnelled over the SSH connection, exposing machines
at either end of the SSH connection to each other over a secure link.
This allows a sort of "poor man's VPN" that can be established using the
existing SSH infrastructure, which is often sufficient to enable a wide
range of use cases. We should support it too.

Modifications:

Wide-ranging. Added the idea of an SSH channel type, added support for
exposing those in the channel initializer callbacks, plumbed through
support for global requests in the form of a global request delegate,
and many more. Substantially reworked a bunch of the API and internals,
basically.

Result:

Remote and local TCP port forwarding work swimmingly well.