Support unknown global request types
Motivation:
SSH would throw errors when an unknown global request was received, and the message would be ignored (in violation of spec).

Solution:
Now, a message of SSH_MSG_REQUEST_FAILURE will be returned to the sender, and the unknown message is passed through far enough that potential future implementations could handle custom global requests.