Add better support for error on receipt of unimplemented (#23)
Motivation:

When we receive the SSH_MSG_UNIMPLEMENTED it's because we sent a message
the remote peer does not understand. We can't do much about this, but we
can construct an error that is clearer about what exactly happened
instead of using a generic protocol error.

Modifications:

- Better erroring on receipt of SSH_MSG_UNIMPLEMENTED.

Result:

Users will be less confused if this happens.