Add basic support for SSH certificates.
Motivation:

Traditionally, SSH public key authentication uses a rudimentary
full-mesh solution to key distribution. If a user wishes to log in to a
machine with a public key (or validate a host key), they need to distribute
their public key ahead of time, out-of-band.

This approach has the advantage of simplicity, making it a great fit for
small-scale operation, with small numbers of users and machines.
However, as organisations grow it begins to represent a limitation:
working out how to distribute keys and associate policy with those keys
becomes a substantial technical challenge.

To address this, SSH has support for a simplified certificate authority
model. This allows servers and clients to delegate their authority to
one or more certificate authorities, which become responsible for
authenticating keys and providing policies for the keyholders.

In larger organisations, this lets servers and clients be configured
with a small number of CA keys, and then to apply policy to the
certificates they receive. Certificate authorities can restrict the
validity of certificates or their capabilities, allowing them to
restrict the authority of keyholders. This provides a powerful policy
engine that is extremely valuable in larger enterprises.

There are two good reasons for swift-nio-ssh to want to build out
support for SSH certificates. The first is a simple compatibility issue:
SSH certificates present themselves as a somewhat specialised type of
public key. In this case, we'd like to be able to work with those kinds
of keys, if in no more complex way than by treating them as
straightforward public keys.

The more important reason is that SSH certificates unlock a bunch of
powerful tools for management of SSH in larger organisations. These
tools are useful to be able to work with programmatically, and so we can
unlock more user value by supporting these key types. This will require
further investment going forward, as this initial patch is somewhat
limited in scope.

Modifications:

This patch adds support for parsing/serialising SSH certificates to the
wire format, as well as parsing of the OpenSSH public key format for SSH
certificates. It also reworks the `NIOSSHPrivateKey` type to be able to
encapsulate a certificate. As part of this work, we also promote
`SSHSignature` to become a public type, renaming it to `NIOSSHSignature`
in the process.

Finally, we add a bunch of tests for the basic functionality, including
validation.

This patch does not contain the following (fairly important)
functionality:

1. No support for generating and signing certificates.
2. No support for serializing the OpenSSH public key format for
   certificates.
3. No proper support for using certificates generated by other tools, as
   we don't support parsing the OpenSSH private key format.

This functionality will come along in later patches.

Results:

The basic SSH certificate infrastructure is in place. This provides us a
jumping off point for implementing the rest of the functionality. In
particular, we are highly unlikely to need new public types from this
point onward: we'd just be extending the types defined here.