Correctly handle host key negotiation.
Motivation:

We took a shortcut in our code for negotiating which host key algorithm
to use, which is that we preferred the first one in our list that was
also in the peer's. This is obviously not right: the preference is
supposed to be the first one in the client's that's also in the
server's. We just happened to be unlucky enough to get away with this
for now.

While I was investigating this I also spotted that we never actually
validated that the server respected the negotiation result, so that code
had to be added too.

Modifications:

- Updated the negotiation code to validate that the server sends back a
  key of the right type.
- Updated the negotiation code to ensure that we negotiate host keys per
  the spec.

Result:

We negotiate host keys the way we're supposed to.