Do not lose data on channel close.
Motivation:

We didn't notice this as much when we had autoRead turned on and the
broken data fast path, but now that we don't it has become apparent that
we can see data loss if CHANNEL_CLOSE arrives while we're queueing
reads. That's not good!

Here we decide to adopt the same pattern as NIO does with TCP channels:
if we receive a hard-close, we will deliver all pending data, even if
you have not got an outstanding read(). This avoids data loss, while
allowing us to rapidly drop resources that are potentially in short
supply. In this case, if the remote peer doesn't care about this channel
anymore, we should really stop holding it open ourselves.

Modifications:

- No longer drop reads in any circumstance, always deliver them.
- Wrote some tests.

Result:

No data loss!