Initialize state machine on handlerAdded (#43)
Motivation:
Sometimes we already have an established channel, for example,
one returned by establishing a tunneled connection. If this
is the case, then we cannot use it to start key exchange, since
it's only started when channel becomes active. In order to
support more use cases we should handle starting key exchange
when handler is added as well.

Modifications:
1. Extract key exchange initialization to separate function
2. Call it from handlerAdded if channel is active
3. Guard againts calling it twice

Result:
Fixes #42