Implement userauth banner processing
Motivation:

SSH servers may choose to send a userauth banner to clients during the
user authentication step of the SSH protocol. The client is obliged to
show this message to the end user, unless explicitely told to ignore
banners.

As of yet, ssh clients built with swift-nio-ssh did not have a proper
way to receive banner messages and could therefore technically not
fullfill the contract laid out in RFC 4252.  Additionally, ssh servers
built with swift-nio-ssh could not send userauth banners to clients,
leaving this part of RFC 4252 entirely unimplemented. For completeness
sake, and to be able to E2E test receiving userauth banners, this patch
implements said behavior.

Modifications:

- Devise e2e test cases to ensure that by default, no banner is sent
  by server
- Devise e2e test case to ensure that banner messages are correctly
  sent and received
- Add new UserAuthBannerEvent that may be processed using
  ChannelInboundHandler's userInboundEventTriggered(...)
- Add package user facing UserAuthBanner type and documentation
- Add new optional 'banner' property to SSHServerConfiguration
- Implement new helper functions in accordance with existing code,
  that permit processing userauth banners on the server while
  being non-invasive to the current architecture
- Implement banner sending at freely choosen, appropriate time right
  before userAuthSuccess. The idea here is to make sure that banners
  will only ever be sent to authorized clients, preventing possible
  unwanted information leaks.

Result:

swift-nio-ssh's implementation of RFC 4252 is more complete, i.e.:

- SSH clients built with swift-nio-ssh are now able to receive and
  correctly process userauth banners
- SSH servers built with swift-nio-ssh are now able to send arbitrary
  userauth banners to clients, via an easily settable SSHServerConfiguration
  property.

Due to the choice of nil as the default value for this property,
existing behavior for swift-nio-ssh servers does not change, i.e.,
userauth banners are explicitely opt-in

Furthermore, e2e tests of userauth banners are now possible