fixes an issue where decodeFirstBlock can be called multiple times (#102)
Motivation:
When we call `decodeFirstBlock` we expect that first blockSize bytes,
that include packet length, will be overwritten in the parser buffer
(this is why we pass source byte buffer as inout). This changes state of
the parser, but if we cannot decode a packet we do not return it and in
the connection state machine we return without updating the machine
state. This can lead to calling decodeFirstBlock multiple times until we
have enough data to decode the whole packet in one go.

Modifications:
 - Update connection state machine state when active even if we do not
   have response message
 - Adds TinyEncriptionAlgorithm implementation to test codepaths that
   use decodeFirstBlock (unlike AESGCM)
 - Test that drip feeds data to parser