Correctly resize ByteBuffers
Motivation:

In two different places in this module we "reserve" space by
moving the writer index forward. This isn't the correct way to
do that, and it's a bit of a needless performance
optimization. Instead, just write a zero in that location and
come back to it.

Modifications:

- Replace unmotivated writer index move with a write.
- Add tests.

Result:

Fewer crashes