Lower alloc limits to lock-in upstream wins
Motivation:

Changes in swift-nio and swift-atomics have resulted in fewer
allocations in the benchmarks in swift-nio-ssh, so much improved that it
caused tests to fail.

Modifications:

Update the limits to take account of the gains, update the script to
scrape the new limits to use 5.8, 5.9, 5.10.

Result:

Allocation benchmark CI should pass again.