Update nightly-next thresholds to lock-in wins
Motivation:

Switching nightly-next to 6.2 nightlies lowered allocations in one of
the benchmarks.

Modifications:

Update the benchmark thresholds. This is the first time it has been done
using a script so some rounding has taken place for some of the
benchmarks but it should be within the tolerances.

Result:

Benchmarks checks for nightly-next pass.